Move reconnect and temporary_disconnect to picard.util.qt
Both helpers operate on Qt signals, so they move to the Qt-only
picard.util.qt. This removes the last remaining PyQt6 users from
picard.util except load_json and encoded_queryargs, which the following
commits rewrite without Qt. Update the importers (ui.logview,
ui.mainwindow, ui.metadatabox.edittagdialog, ui.player.listenbrainz,
ui.widgets.pluginlistwidget).
