+++ picard/util/qt.py
@@ -28,10 +28,12 @@ tooling or a potential headless/alternative frontend).
 
 from collections.abc import (
     Callable,
+    Generator,
     Iterable,
     Iterator,
     Mapping,
 )
+from contextlib import contextmanager
 from time import monotonic
 from typing import Any
 
@@ -174,3 +176,39 @@ def build_qurl(
 def parse_json(reply: QNetworkReply) -> Any:
     """Deserialize the JSON body of a ``QNetworkReply`` to a Python object."""
     return load_json(reply.readAll().data())
+
+
+def reconnect(
+    signal: QtCore.pyqtBoundSignal, newhandler: Callable | None = None, oldhandler: Callable | None = None
+) -> None:
+    """
+    Reconnect an handler to a signal
+
+    It disconnects all previous handlers before connecting new one
+
+    Credits: https://stackoverflow.com/a/21589403
+    """
+    while True:
+        try:
+            if oldhandler is not None:
+                signal.disconnect(oldhandler)
+            else:
+                signal.disconnect()
+        except TypeError:
+            break
+    if newhandler is not None:
+        signal.connect(newhandler)
+
+
+@contextmanager
+def temporary_disconnect(signal: QtCore.pyqtBoundSignal, *handlers: Callable) -> Generator[None, None, None]:
+    """
+    Create context to temporarly disconnect one or more signal handlers
+    """
+    try:
+        for handler in handlers:
+            signal.disconnect(handler)
+        yield
+    finally:
+        for handler in handlers:
+            signal.connect(handler)
